ctf-web-app/requirements.txt
Django==1.6.1
Markdown==2.3.1
South==0.8.4
django-haystack==2.1.0
django-model-utils==1.5.0
djangorestframework==2.3.10
psycopg2==2.5.1
pyelasticsearch==0.6.1
pytz==2013.8
requests==2.2.1
shortuuid==0.4
simplejson==3.3.3
six==1.5.2
urllib3==1.7.1
wsgiref==0.1.2
